pyproject.toml
[project]
name = "leafmine-signature"
version = "0.1.0"
description = "Asset pipeline for signature analysis on leaf mine"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.49.1",
    "numpy>=2.1.0",
    "Pillow>=10.4.0",
    "roughpy",
    "scikit-image>=0.24.0",
    "ruff>=0.14.1,<0.15",
]

[project.scripts]
gradio_app = "views.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["views", "models", "controllers"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["controllers", "models", "views"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
